Refactor for re-useability
