+++ cypress/utils/composer/createArticle.js
@@ -1,0 +1,8 @@
+import {wait} from "../wait";
+
+export function createArticle() {
+    cy.get("#js-dashboard-create-dropdown").click();
+    cy.get("#js-dashboard-create-article").click();
+    wait(2);
+}
+
