+++ cypress/utils/composer/startEditing.js
@@ -1,0 +1,7 @@
+import { wait } from "../wait";
+
+export function startEditing() {
+    cy.get(".ProseMirror").click();
+    wait(1);
+}
+
