+++ cypress/utils/wait.js
@@ -1,1 +1,1 @@
-export function wait(seconds) { cy.wait(seconds * 1000); }
+export function wait(seconds) { return cy.wait(seconds * 1000); }
